+++ api/v2/names.go
@@ -1,4 +1,4 @@
-package common
+package v2
 
 import (
 	"fmt"
