+++ api/v2/names_test.go
@@ -1,4 +1,4 @@
-package common
+package v2
 
 import (
 	"testing"
